+++ console2/src/components/organisms/index.ts
@@ -27,6 +27,7 @@ export { default as NewProjectActivity } from './NewProjectActivity';
 export { default as NewSecretActivity } from './NewSecretActivity';
 export { default as NewTeamActivity } from './NewTeamActivity';
 export { default as OrganizationList } from './OrganizationList';
+export { default as ProcessActivity } from './ProcessActivity';
 export { default as ProcessFormActivity } from './ProcessFormActivity';
 export { default as ProcessList } from './ProcessList';
 export { default as ProcessLogActivity } from './ProcessLogActivity';